requirement.txt
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.1
paramiko==3.4.1
Pillow==9.2.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
